Fix Bundle().toExtras()
